MCP/mcp-postgress-audio/pyproject.toml
[project]
name = "mcp-postgress-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
